+++ docs/releasenotes/releasenotes.rst
@@ -8,6 +8,16 @@ For Dashticz's **master** version Release Notes go to: https://dashticz.readthed
 Recent changes
 ---------------
 
+v3.8.2 Beta (24-4-2021)
+-----------------------
+
+.. note:: Breaking changes: New weather block.
+
+Enhancements
+~~~~~~~~~~~~
+
+* Rewrite of the weather block. See :ref:`customweather`.
+
 Fixes
 ~~~~~~